MCP Client (React & Electron)/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.65.0",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]
